agent_engine/langgraph_basic/requirements.txt
google-cloud-aiplatform==1.110.0
langchain-google-vertexai==2.0.28
langchain-tavily==0.2.11
langgraph==0.3.34
langsmith==0.4.15
python-dotenv==1.1.1
typing-extensions==4.14.1
langchain-core==0.3.74
langchain==0.3.27
openinference-instrumentation-langchain==0.1.34